+++ snowflake/ml/model/_packager/model_meta/model_meta_schema.py
@@ -82,6 +82,7 @@ class SentenceTransformersModelBlobOptions(BaseModelBlobOptions):
 
 ModelBlobOptions = Union[
     BaseModelBlobOptions,
+    CatBoostModelBlobOptions,
     HuggingFacePipelineModelBlobOptions,
     MLFlowModelBlobOptions,
     XgboostModelBlobOptions,